workflow/requirements.txt
chromadb==1.0.17
gradio==5.43.1
gradio_client==1.12.1
newsapi-python==0.2.7
openai==1.99.9
openai-agents==0.2.6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
sendgrid==6.12.4
uv==0.8.12
